+++ src/scripts/parts.js
@@ -10,7 +10,7 @@ import ProjectsList from './react/layout/project-select/List';
 import NewProjectModal from './react/layout/NewProjectModal';
 import UserLinks from './react/layout/UserLinks';
 import ProjectOverview from './react/admin/project-graph/Index';
-import RegionSelect from './react/login/RegionSelect';
+import RegionSelect from './react/admin/login/RegionSelect';
 
 export default {
   helpers,
